+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# [1.7.0-next.2](https://github.com/SaulMoro/ngrx-rtk-query/compare/v1.7.0-next.1...v1.7.0-next.2) (2021-04-19)
+
+
+### Features
+
+* **lib:** add customHandler to module config ([ad69727](https://github.com/SaulMoro/ngrx-rtk-query/commit/ad6972731d80c07948849f786b82d456e38a10f4))
+* **lib:** uadd mutation selectFromResult & tags ([6b2f71c](https://github.com/SaulMoro/ngrx-rtk-query/commit/6b2f71cc348f5d902f26f0fbf85870770046d576))
+
+# [1.7.0-next.1](https://github.com/SaulMoro/ngrx-rtk-query/compare/v1.6.2-next.1...v1.7.0-next.1) (2021-04-16)
+
+
+### Features
+
+* **lib:** refactor thunk svr for better perf ([5cf206b](https://github.com/SaulMoro/ngrx-rtk-query/commit/5cf206b455aa15df8af6f804b3e2af6aa621ec53))
+
 ## [1.6.2](https://github.com/SaulMoro/ngrx-rtk-query/compare/v1.6.1...v1.6.2) (2021-04-09)